+++ CHANGES.rst
@@ -5,6 +5,8 @@ These features will be included in the next release:
 
 Added
 -----
+- If AST differs with zero context lines, search for the lowest successful number of
+  context lines using a binary search to improve performance
 
 Fixed
 -----